FT: upgrade s3cmd to v1.6.1 for s3 ci runs
In previous version (v1.6.0) s3cmd info returns NotImplemented error if some of the queries are not supported. We prefer customers to use the newer version which will print none for any fields that receive NotImplemented error and still displays output for any query requests that succeed (with the exception if ?location query requests are not supported).